@@ -47,7 +47,7 @@
     "directline",
     "automated",
     "testing",
-    "chatdown",
+    "emulator",
     "transcript",
     "chat"
   ],
@@ -63,29 +63,28 @@
     "@types/chalk": "^2.2.0",
     "@types/mocha": "^5.2.6",
     "@types/npmlog": "^4.1.1",
-    "@types/sinon": "^7.0.11",
-    "@types/node": "^11.13.7",
-    "@types/request": "^2.48.1",
-    "@types/request-promise": "^4.1.42",
+    "@types/sinon": "^7.0.13",
+    "@types/node": "^12.7.3",
+    "@types/request": "^2.48.2",
+    "@types/request-promise": "^4.1.44",
     "chai": "^4.2.0",
     "chai-spies": "^1.0.0",
     "coveralls": "^3.0.3",
     "istanbul": "^0.4.5",
-    "mocha": "^6.1.4",
+    "mocha": "^6.2.0",
     "nock": "^10.0.6",
     "nyc": "^14.1.1",
     "proxyquire": "^2.1.0",
-    "sinon": "^7.3.2",
+    "sinon": "^7.4.2",
     "source-map-support": "^0.5.12",
-    "ts-node": "^8.1.0",
-    "ts-node-dev": "^1.0.0-pre.32",
-    "typescript": "^3.4.4",
+    "ts-node": "^8.3.0",
+    "ts-node-dev": "^1.0.0-pre.42",
+    "typescript": "^3.6.2",
     "reflect-metadata": "^0.1.13"
   },
   "dependencies": {
     "chalk": "^2.4.2",
-    "chatdown": "^1.2.0",
-    "commander": "^2.20.0",
+    "commander": "^3.0.1",
     "npmlog": "^4.1.2",
     "request": "^2.88.0",
     "request-promise": "^4.2.4"
