@@ -3,11 +3,10 @@ import { FileInfo } from './domain/fileInfo';
 import { Extensions, ActivityTypes } from './constants';
 import { JabberActivity } from './domain/jabberActivity';
 import fs = require('fs');
-import * as chatdown from 'chatdown';
-import log = require('npmlog');
+import { Activity } from './domain/activity';
 
 /**
- * Transcript generator which consumes a .chat file to generate mock transcripts. Uses existing tool "chatdown" (https://github.com/microsoft/botbuilder-tools/tree/master/packages/Chatdown)
+ * Transcript generator which consumes a .transcript file to generate mock transcripts.
  */
 export class TranscriptGenerator {
 
@@ -19,22 +18,11 @@ export class TranscriptGenerator {
     * @return {Promise<Activity[]} A promise with an array of activities that have been generated from the file
     */
 
-  async single(file: FileInfo, isAzureDevopsTask: boolean): Promise<chatdown.Activity[]> {
+  async single(file: FileInfo, isAzureDevopsTask: boolean): Promise<Activity[]> {
     var fileContents = fs.readFileSync(path.resolve(file.path), 'utf8');
-    var activities: chatdown.Activity[] = new Array<chatdown.Activity>();
+    var activities: Activity[] = new Array<Activity>();
 
-    if (file.extension == Extensions.chatdown) {
-      if (isAzureDevopsTask) {
-        log.warn("WRN", `Chatdown is not supported using azure devops pipeline tasks. Please convert to a transcript file first.`)
-      }
-      else {
-        var args = { in: file.path };
-        await chatdown.default(fileContents, args).then((fileActivities: chatdown.Activity[]) => {
-          activities = fileActivities.filter(x => x.from && x.recipient);
-        });
-      }
-    }
-    else if (file.extension == Extensions.transcript) {
+    if (file.extension == Extensions.transcript) {
       var jsonActivities = JSON.parse(fileContents);
 
       if (Array.isArray(jsonActivities)) {
