@@ -13,20 +13,20 @@ class App {
 
         commander
             .version('1.0.0')
-            .description('a tool that communicates over directline that will test your Chatdown conversations')
+            .description('a tool that communicates over directline that will test your bot conversations')
 
         commander
             .command('test')
-            .option('-f, --files [files]', 'chatdown file(s) to test. Add multiple by comma seperating them.', commaSeparatedList)
-            .option('-d, --dirs [dirs]', 'directory or directories with chatdown files to test', commaSeparatedList)
+            .option('-f, --files [files]', 'transcript file(s) to test. Add multiple by comma seperating them.', commaSeparatedList)
+            .option('-d, --dirs [dirs]', 'directory or directories with transcript files to test', commaSeparatedList)
             .option('-s, --secret [secret]', 'directline secret for authentication', null)
             .option('-e, --endpoint [endpoint]', 'endpoint to retrieve directline token', null)
-            .option('-r, --recursive', 'includes subfolders when scanning for chatdown files', true)
+            .option('-r, --recursive', 'includes subfolders when scanning for transcript files', true)
             .option('-v, --verbose', 'enables verbose logging', false)
             .description('test a scenario using directline')
             .action(async function(options: any) {
                 if ((!options.files || options.files.length <= 0) && (!options.dirs || options.dirs.length <= 0)) {
-                    log.error('ERR', 'please specify chatdown file(s) (--files) and/or a directory (--dirs) that contains your chatdown files');
+                    log.error('ERR', 'please specify transcript file(s) (--files) and/or a directory (--dirs) that contains your transcript files');
                     return;
                 }
 
